move files from state to model
