+++ report-viewer/src/components/fileDisplaying/CodePanel.vue
@@ -69,7 +69,7 @@ import type { MatchInSingleFile } from '@/model/MatchInSingleFile'
 import { ref, nextTick, type PropType, computed, type Ref } from 'vue'
 import Interactable from '../InteractableComponent.vue'
 import type { Match } from '@/model/Match'
-import type { SubmissionFile } from '@/stores/state'
+import type { SubmissionFile } from '@/model/File'
 import { highlight } from '@/utils/CodeHighlighter'
 import type { ParserLanguage } from '@/model/Language'
 import { getMatchColor } from '@/utils/ColorUtils'
