+++ report-viewer/src/model/fileHandling/FileHandler.ts
