+++ report-viewer/src/model/ui/DistributionChartConfig.ts
@@ -1,0 +1,9 @@
+import type { MetricType } from '../MetricType'
+
+/**
+ * Configuration for the distribution chart.
+ */
+export interface DistributionChartConfig {
+  metric: MetricType
+  xScale: 'linear' | 'logarithmic'
+}
