+++ report-viewer/src/views/FileUploadView.vue
@@ -57,8 +57,8 @@ import { store } from '@/stores/store'
 import Button from '@/components/ButtonComponent.vue'
 import VersionInfoComponent from '@/components/VersionInfoComponent.vue'
 import LoadingCircle from '@/components/LoadingCircle.vue'
-import { JsonFileHandler } from '@/utils/fileHandling/JsonFileHandler'
-import { ZipFileHandler } from '@/utils/fileHandling/ZipFileHandler'
+import { ZipFileHandler } from '@/model/fileHandling/ZipFileHandler'
+import { JsonFileHandler } from '@/model/fileHandling/JsonFileHandler'
 
 store().clearStore()
